requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
astroid==2.4.2
dataclasses==0.8
Django==3.0.5
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
promise==2.3
psycopg2==2.8.6
pylint==2.6.0
python-decouple==3.4
pytz==2021.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.2.4
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
wrapt==1.12.1
